# api-spec.yaml
openapi: 3.1.0
info:
  title: MS Template API
  description: "Microservice template API exposing company, salesman, admin user,\
    \ and external posts endpoints. All endpoints support request context propagation\
    \ using X-Request-Id, X-Tenant-Id, and Accept-Language headers."
  contact:
    name: MS Template Team
    url: https://snb.com/support
    email: support@snb.com
  license:
    name: Proprietary
    url: https://snb.com/legal
  version: v1
...
servers:
- url: http://localhost:8080
  description: Local development
- url: /
  description: Relative deployment base path
...
paths:
...
  /api/rbac/roles:
    get:
      tags:
      - RBAC
      summary: List roles
      description: Returns all active role records.
      operationId: getAllRoles
      parameters:
      - $ref: "#/components/parameters/XRequestIdHeader"
      - $ref: "#/components/parameters/XTenantIdHeader"
      - $ref: "#/components/parameters/AcceptLanguageHeader"
      responses:
        "200":
          description: Roles fetched successfully
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/RoleResponse"
              examples:
                RolesList:
                  description: RolesList
                  value:
                  - roleId: 1
                    roleCode: SUPER_ADMIN
                    roleName: Super Admin
                    description: Full platform access
                    createdBy: 1001
                    createdAt: 2026-05-04T10:15:30
                    updatedBy: 1002
                    updatedAt: 2026-05-04T11:20:45
                  - roleId: 2
                    roleCode: AUDITOR
                    roleName: Auditor
                    description: Read-only audit access
                    createdBy: 1001
                    createdAt: 2026-05-04T10:30:00
                    updatedBy: 1002
                    updatedAt: 2026-05-04T11:30:00
        "500":
          description: Internal server error
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/BaseResponseDTO"
    post:
      tags:
      - RBAC
      summary: Create role
      description: Creates a new role record.
      operationId: createRole
      parameters:
      - $ref: "#/components/parameters/XRequestIdHeader"
      - $ref: "#/components/parameters/XTenantIdHeader"
      - $ref: "#/components/parameters/AcceptLanguageHeader"
      requestBody:
        description: Role payload to create
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RoleCreateRequest"
            examples:
              CreateRole:
                description: CreateRole
                value:
                  roleCode: SUPER_ADMIN
                  roleName: Super Admin
                  description: Full platform access
        required: true
      responses:
        "201":
          description: Role created successfully
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/RoleResponse"
              examples:
                CreatedRole:
                  description: CreatedRole
                  value:
                    roleId: 1
                    roleCode: SUPER_ADMIN
                    roleName: Super Admin
                    description: Full platform access
                    createdBy: 1001
                    createdAt: 2026-05-04T10:15:30
                    updatedBy: 1001
                    updatedAt: 2026-05-04T10:15:30
        "400":
          description: Validation failed
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/BaseResponseDTO"
        "500":
          description: Internal server error
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/BaseResponseDTO"
  /api/rbac/roles/bulk:
    post:
      tags:
      - RBAC
      summary: Create roles in bulk
      description: Creates multiple role records in a single request.
      operationId: createRolesBulk
      parameters:
      - $ref: "#/components/parameters/XRequestIdHeader"
      - $ref: "#/components/parameters/XTenantIdHeader"
      - $ref: "#/components/parameters/AcceptLanguageHeader"
      requestBody:
        description: Bulk role payload to create
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RoleBulkCreateRequest"
            examples:
              CreateRolesBulk:
                description: CreateRolesBulk
                value:
                  roles:
                  - roleCode: SUPER_ADMIN
                    roleName: Super Admin
                    description: Full platform access
                  - roleCode: AUDITOR
                    roleName: Auditor
                    description: Read-only audit access
        required: true
      responses:
        "201":
          description: Roles created successfully
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/RoleBulkResponse"
              examples:
                CreatedRolesBulk:
                  description: CreatedRolesBulk
                  value:
                    roles:
                    - roleId: 1
                      roleCode: SUPER_ADMIN
                      roleName: Super Admin
                      description: Full platform access
                      createdBy: 1001
                      createdAt: 2026-05-04T10:15:30
                      updatedBy: 1001
                      updatedAt: 2026-05-04T10:15:30
                    - roleId: 2
                      roleCode: AUDITOR
                      roleName: Auditor
                      description: Read-only audit access
                      createdBy: 1001
                      createdAt: 2026-05-04T10:15:31
                      updatedBy: 1001
                      updatedAt: 2026-05-04T10:15:31
        "400":
          description: Validation failed
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/BaseResponseDTO"
        "500":
          description: Internal server error
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/BaseResponseDTO"
  /api/rbac/roles/{roleCode}:
    get:
      tags:
      - RBAC
      summary: Get role by code
      description: Finds a role by code.
      operationId: getRoleByCode
      parameters:
      - name: roleCode
        in: path
        description: Role code
        required: true
        schema:
          type: string
        example: SUPER_ADMIN
      - $ref: "#/components/parameters/XRequestIdHeader"
      - $ref: "#/components/parameters/XTenantIdHeader"
      - $ref: "#/components/parameters/AcceptLanguageHeader"
      responses:
        "200":
          description: Role fetched successfully
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/RoleResponse"
              examples:
                RoleByCode:
                  description: RoleByCode
                  value:
                    roleId: 1
                    roleCode: SUPER_ADMIN
                    roleName: Super Admin
                    description: Full platform access
                    createdBy: 1001
                    createdAt: 2026-05-04T10:15:30
                    updatedBy: 1002
                    updatedAt: 2026-05-04T11:20:45
        "400":
          description: Invalid role code supplied
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/BaseResponseDTO"
        "404":
          description: Role not found
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/BaseResponseDTO"
        "500":
          description: Internal server error
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/BaseResponseDTO"
    put:
      tags:
      - RBAC
      summary: Update role
      description: Updates an existing role record by code.
      operationId: updateRole
      parameters:
      - name: roleCode
        in: path
        description: Role code
        required: true
        schema:
          type: string
        example: SUPER_ADMIN
      - $ref: "#/components/parameters/XRequestIdHeader"
      - $ref: "#/components/parameters/XTenantIdHeader"
      - $ref: "#/components/parameters/AcceptLanguageHeader"
      requestBody:
        description: Role payload to update
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RoleUpdateRequest"
            examples:
              UpdateRole:
                description: UpdateRole
                value:
                  roleName: Super Administrator
                  description: Updated description
        required: true
      responses:
        "200":
          description: Role updated successfully
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/RoleResponse"
              examples:
                UpdatedRole:
                  description: UpdatedRole
                  value:
                    roleId: 1
                    roleCode: SUPER_ADMIN
                    roleName: Super Administrator
                    description: Updated description
                    createdBy: 1001
                    createdAt: 2026-05-04T10:15:30
                    updatedBy: 1002
                    updatedAt: 2026-05-04T12:15:30
        "400":
          description: Validation failed
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/BaseResponseDTO"
        "404":
          description: Role not found
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/BaseResponseDTO"
        "500":
          description: Internal server error
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/BaseResponseDTO"
    delete:
      tags:
      - RBAC
      summary: Delete role
      description: Soft-deletes a role record by code.
      operationId: deleteRole
      parameters:
      - name: roleCode
        in: path
        description: Role code
        required: true
        schema:
          type: string
        example: SUPER_ADMIN
      - $ref: "#/components/parameters/XRequestIdHeader"
      - $ref: "#/components/parameters/XTenantIdHeader"
      - $ref: "#/components/parameters/AcceptLanguageHeader"
      responses:
        "200":
          description: Role deleted successfully
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/RoleResponse"
              examples:
                DeletedRole:
                  description: DeletedRole
                  value:
                    roleId: 1
                    roleCode: SUPER_ADMIN
                    roleName: Super Administrator
                    description: Updated description
                    createdBy: 1001
                    createdAt: 2026-05-04T10:15:30
                    updatedBy: 1002
                    updatedAt: 2026-05-04T12:20:00
        "404":
          description: Role not found
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/BaseResponseDTO"
        "500":
          description: Internal server error
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/BaseResponseDTO"
...
components:
  schemas:
    RoleCreateRequest:
      type: object
      description: Role creation request payload
      properties:
        roleCode:
          type: string
          description: Stable role code used in authorization checks
          example: SUPER_ADMIN
          maxLength: 100
          minLength: 1
        roleName:
          type: string
          description: Human-readable role name
          example: Super Admin
          maxLength: 150
          minLength: 1
        description:
          type: string
          description: Role description
          example: Full platform access
          maxLength: 255
      required:
      - roleCode
      - roleName
    RoleUpdateRequest:
      type: object
      description: Role update request payload
      properties:
        roleName:
          type: string
          description: Human-readable role name
          example: Super Administrator
          maxLength: 150
          minLength: 1
        description:
          type: string
          description: Role description
          example: Updated description
          maxLength: 255
      required:
      - roleName
    RoleBulkCreateRequest:
      type: object
      description: Bulk role creation request payload
      properties:
        roles:
          type: array
          description: Role payload list to create
          items:
            $ref: "#/components/schemas/RoleCreateRequest"
      required:
      - roles
    RoleBulkResponse:
      type: object
      description: Bulk role creation response payload
      properties:
        roles:
          type: array
          description: Created role records
          items:
            $ref: "#/components/schemas/RoleResponse"
      required:
      - roles
    RoleResponse:
      type: object
      description: Role response payload
      properties:
        roleId:
          type: integer
          format: int64
          description: Unique role identifier
          example: 1
        roleCode:
          type: string
          description: Stable role code used in authorization checks
          example: SUPER_ADMIN
        roleName:
          type: string
          description: Human-readable role name
          example: Super Admin
        description:
          type: string
          description: Role description
          example: Full platform access
        createdBy:
          type: integer
          format: int64
          description: Identifier of actor who created the record
          example: 1001
        createdAt:
          type: string
          format: date-time
          description: Record creation timestamp in ISO-8601 format
          example: 2026-05-04T10:15:30
        updatedBy:
          type: integer
          format: int64
          description: Identifier of actor who last updated the record
          example: 1002
        updatedAt:
          type: string
          format: date-time
          description: Last update timestamp in ISO-8601 format
          example: 2026-05-04T11:20:45
...
    ErrorInfo:
      type: object
      description: Detailed error information
      properties:
        type:
          type: string
          description: Error category
          example: VALIDATION_ERROR
        code:
          type: string
          description: Application-specific error code
          example: EMAILADDRESS_INVALID
        message:
          type: string
          description: "Localized, user-friendly error message"
          example: emailAddress must be a valid email
        description:
          type: string
          description: Additional diagnostic details
          example: "Rejected value: not-an-email"
    BaseResponseDTO:
      type: object
      description: Base API response wrapper with error details
      properties:
        errors:
          type: array
          items:
            $ref: "#/components/schemas/ErrorInfo"
...
  parameters:
    XRequestIdHeader:
      name: X-Request-Id
      in: header
      description: "Correlation identifier for distributed tracing. If omitted, the\
        \ service generates one."
      required: false
      schema:
        type: string
        example: req-20260504-0001
    XTenantIdHeader:
      name: X-Tenant-Id
      in: header
      description: Tenant identifier used for multi-tenant context propagation.
      required: false
      schema:
        type: string
        example: tenant-001
    AcceptLanguageHeader:
      name: Accept-Language
      in: header
      description: Preferred locale for localized validation and error messages.
      required: false
      schema:
        type: string
        example: en